// types/ad.ts
export type Ad = {
  id: number;
  mainDirectionId: number;
  directionId: number;
  subDirectionId: number;
  categoryId: number;
  provinceId: number;
  districtId: number;
  khorooId: number;
  title: string;
  address: string;
  desciption: string;
  price: number;
  counter: number;
  email: string;
  phone: string;
  isMessenger: boolean;
  isTermOfService: boolean;
  createdBy?: string;
  createdAt?: string;
  updatedBy?: string;
  updatedAt?: string;
  deletedAt?: string;
  province: {
    id: number;
    name: string;
    parentId: number;
    position: number;
    type: string;
    createdAt?: string;
    updatedAt?: string;
    deletedAt?: string;
  };
  district: {
    id: number;
    name: string;
    parentId: number;
    position: number;
    type: string;
    createdAt?: string;
    updatedAt?: string;
    deletedAt?: string;
  };
  khoroo: {
    id: number;
    name: string;
    parentId: number;
    position: number;
    type: string;
    createdAt?: string;
    updatedAt?: string;
    deletedAt?: string;
  };
  images?: string;
};